+++ packages/subtree-updater/build_docker.sh
@@ -22,7 +22,7 @@ fi
 
 if [ ! -z "$IS_MOCK" ]; then
     echo "building mock subtree-updater"
-    docker build -t nocturnexyz/mock-subtree-updater:dev -f ./packages/subtree-updater/Mock.Dockerfile .
+    docker build -t nocturnelabs/mock-subtree-updater:dev -f ./packages/subtree-updater/Mock.Dockerfile .
     exit 0
 fi
 
@@ -31,9 +31,9 @@ if [[ $(uname -m) == 'arm64' ]]; then
 	echo "dected arm64, building for x86 using docker buildx..."
 
     docker buildx build --platform linux/amd64 -t rapidsnark ./rapidsnark
-    docker buildx build --platform linux/amd64 -t nocturnexyz/subtree-updater:dev -f ./packages/subtree-updater/Dockerfile .
+    docker buildx build --platform linux/amd64 -t nocturnelabs/subtree-updater:dev -f ./packages/subtree-updater/Dockerfile .
 else
     echo 
     docker build -t rapidsnark ./rapidsnark
-    docker build -t nocturnexyz/subtree-updater:dev -f ./packages/subtree-updater/Dockerfile .
+    docker build -t nocturnelabs/subtree-updater:dev -f ./packages/subtree-updater/Dockerfile .
 fi