+++ scripts/push-docker-dev.sh
@@ -1,0 +1,3 @@
+docker push nocturnelabs/bundler:dev
+docker push nocturnelabs/deposit-screener:dev
+docker push nocturnelabs/subtree-updater:dev